feat(mfa): add back button on verify identity page
Provide a way for a user to go back/cancel so that they can clear their
session. Otherwise, they would need to dig into their browser settings
to delete the cookie and/or fallback header.

This only adds the button on desktop because the button doesn't fit
right on mobile.