fix(migrations): fix modal not closing
We should be using a bound show variable to control the modal. Prior to
this PR, we were indirectly controlling the modal by setting the
migration ID via the on close handler, but that event is never triggered
when the modal is closed.

This also refactors the modal to pass in the migration instead of a
list of migrations and a migration ID because the entire list is not
needed inside the modal.