fix(functions): convert cron expression to timestamp
In a previous version of Appwrite, there was a next attribute with the
timestamp of when the function would be executed next. That was removed,
but the Console has not been updated to show the correct timestamp.

This PR installs cron-parser package and uses it to conver the schedule
to a human readable timestamp.